+++ Cargo.toml
@@ -24,10 +24,10 @@ egui-winit = { version = "0.31", default-features = false }
 egui = "0.31"
 image = { version = "0.25", optional = true }
 winit = { version = "0.30", default-features = false, features = ["rwh_06"] }
-vulkano = { version = "0.35" }
-vulkano-shaders = "0.35"
+vulkano =           { git = "https://github.com/vulkano-rs/vulkano.git", branch = "master" }
+vulkano-shaders =   { git = "https://github.com/vulkano-rs/vulkano.git", branch = "master" }
+vulkano-taskgraph = { git = "https://github.com/vulkano-rs/vulkano.git", branch = "master" }
 
 [dev-dependencies]
 cgmath = "0.18.0"
 egui_demo_lib = "0.31.0"
-vulkano-util = { version = "0.35" }
